codematics/requirements.txt
djangorestframework
pyyaml
django-cors-headers
django-phonenumber-field[phonenumbers]
Pillow
gunicorn
Django==4.1.4
markdown 
django-filter
stripe
psycopg2-binary
djangorestframework-simplejwt
PyJWT==2.1.0
django-paypal
django-taggit
nanoid
pytest-django
django-dotenv
drf-access-policy
djangorestframework-word-filter
usps-api
django-redis
django-notifications-hq
django-model-utils
urbanairship
ua-parser 
user-agents
django-user-agents
django_cron
